# api_doc.yaml
swagger: "2.0"
info:
  description: "Homework description"
  version: "1.0.0"
  title: "LAB FE 14 NodeJS HW #3"
  termsOfService: "http://swagger.io/terms/"
  contact:
    email: "Kyrylo_Yezhov@epam.com"
host: "localhost:8080"
basePath: "/"
...
schemes:
- "http"
paths:
...
  /api/auth/forgot_password:
    post:
      tags:
      - "Auth"
      summary: "Forgot password"
      description: "(Optional) Forgot password option"
      operationId: "forgotPassword"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "credentials"
        schema:
          properties:
            email:
              type: "string"
              example: "kyrylo@gmail.com"
      responses:
        "200":
            description: "Success"
            schema:
                type: "object"
                properties:
                  message: 
                    type: string
                    example: "New password sent to your email address"
        "400":
          description: "Bad request"
          schema:
            $ref: '#/definitions/Error'
        "500":
          description: "Internal server error"
          schema:
            $ref: '#/definitions/Error'
...
securityDefinitions:
  jwt_token:
    type: "apiKey"
    name: "Authorization"
    in: "header"
definitions:
  Error:
    type: "object"
    properties:
      message:
        type: "string"
        description: "Error message"
...